ofproto: fix OF1.0 packet_in parser
OF1.0 spec says OFP_ASSERT(sizeof(struct ofp_packet_in) == 20).
It's quite bogus as it assumes a specific class of C implementations.
(well, if it was C.  it's unclear from the spec itself.)
We just use the real size of the structure as this is not C.  This
agrees with on-wire messages OpenFlow Reference Release and Open vSwitch
produce.

This should fix a crash Chen Chen reported on ryu-devel recently.

Signed-off-by: YAMAMOTO Takashi <yamamoto@valinux.co.jp>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>