topology: Make sure the last IP in the list is the newest IP.
In the wireless network, when host attaches AP, it will send many ARP
requests which Src IP were used before.  This causes the last IP in
the list may not be the newest one and we also cannot know which IP in
the list is the newest.

So this patch can solve it.

Signed-off-by: Chia-Lin Cho <fox91119@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>