simple_switch: Separate simple_switch for beginners
simple_switch_* apps have conflicting two faces of its purpose.
One is the example apps for beginners, the other is the utility
apps for the L2 switching in the user environments.
These are difficult to be compatible.

This patch adds an example app for begginers and separates
from simple_switch_* apps.
This makes it easy to improve the L2 switching function of
simple_switch_* apps.

To simplify example_switch_13, the following features are dropped
from simple_switch_* apps.
 - FlowMod and Packet-Out messages using "buffer_id".
 - Checking the expected "miss_send_length" field.
 - Ignoring the LLDP packets

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>